requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
certifi==2019.6.16
chardet==3.0.4
cssmin==0.2.0
Django==1.11.5
django-activity-stream==0.6.4
django-appconf==1.0.3
django-autoadmin==1.1.1
django-celery==3.2.1
django-colorful==1.2
django-compressor==2.2
django-cors-headers==2.1.0
django-downloadview==1.9
django-formtools==2.0
django-mathfilters==0.4.0
django-model-utils==3.0.0
django-mptt==0.8.7
django-pure-pagination==0.3.0
django-qsstats-magic==1.0.0
django-rest-swagger==0.3.10
django-solo==1.1.3
django-stronghold==0.2.9
django-suit==0.2.25
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
djangorestframework-recursive==0.1.1
elasticsearch==7.0.2
funcsigs==1.0.2
furl==1.0.1
fusepy==2.0.4
graphviz==0.8
idna==2.6
kombu==3.0.37
mock==2.0.0
olefile==0.46
orderedmultidict==1.0.1
pbr==5.4.0
Pillow==4.2.1
pkg-resources==0.0.0
pycountry==17.5.14
pyocr==0.4.7
PyPDF2==1.26.0
python-dateutil==2.6.1
python-gnupg==0.4.1
python-magic==0.4.13
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
sh==1.12.14
six==1.12.0
urllib3==1.22
